+++ src/resources/chat/completions.ts
@@ -704,14 +704,24 @@ export interface CompletionCreateParamsBase {
     | 'QwQ-32B-Preview'
     | 'Meta-Llama-Guard-3-8B'
     | 'DeepSeek-R1'
+    | 'DeepSeek-R1-0528'
     | 'DeepSeek-V3-0324'
     | 'DeepSeek-V3.1'
+    | 'DeepSeek-V3.1-Terminus'
     | 'DeepSeek-R1-Distill-Llama-70B'
     | 'Llama-4-Maverick-17B-128E-Instruct'
     | 'Llama-4-Scout-17B-16E-Instruct'
     | 'Qwen3-32B'
     | 'Llama-3.3-Swallow-70B-Instruct-v0.4'
-    | 'gpt-oss-120b';
+    | 'gpt-oss-120b'
+    | 'ALLaM-7B-Instruct-preview';
+
+  /**
+   * A dictionary of additional keyword arguments to pass into the chat template. Use
+   * this to provide extra context or parameters that the model's chat template can
+   * process. Keys must be strings; values may be any valid JSON type.
+   */
+  chat_template_kwargs?: CompletionCreateParams.ChatTemplateKwargs | null;
 
   /**
    * If true, sampling is enabled during output generation. If false, deterministic
@@ -1094,6 +1104,21 @@ export namespace CompletionCreateParams {
 
       [k: string]: unknown;
     }
+  }
+
+  /**
+   * A dictionary of additional keyword arguments to pass into the chat template. Use
+   * this to provide extra context or parameters that the model's chat template can
+   * process. Keys must be strings; values may be any valid JSON type.
+   */
+  export interface ChatTemplateKwargs {
+    /**
+     * Enables the model's internal reasoning or "thinking" mode, if supported by the
+     * chat template (deepseek models).
+     */
+    enable_thinking?: boolean;
+
+    [k: string]: unknown;
   }
 
   /**